+++ test/v3_test.dart
@@ -3,6 +3,7 @@ import 'dart:async';
 import 'package:async/async.dart';
 import 'package:postgres/messages.dart';
 import 'package:postgres/postgres.dart';
+import 'package:postgres/src/v3/connection.dart';
 import 'package:stream_channel/stream_channel.dart';
 import 'package:test/test.dart';
 
@@ -25,6 +26,12 @@ void main() {
     });
 
     tearDown(() => connection.close());
+
+    test('runtime parameters', () async {
+      final c = connection as PgConnectionImplementation;
+      final p = c.runtimeParameters;
+      expect(p.applicationName, 'test_app');
+    });
 
     test('simple queries', () async {
       final rs = await connection.execute("SELECT 'dart', 42, NULL");