+++ CHANGELOG.md
@@ -7,6 +7,7 @@
   - (internal) using `EncodedValue` instead of `EncodeOutput`
 - `TypeCodec` interface (may become public) is used for encoding/decoding value by OIDs.
 - `TypeEncoderFn` type definition for generic Dart -> Postgres object encoders (where type is not specified as parameter).
+- `RuntimeParameters` (accessible through `TypeCodecContext`) to access server-provided parameter status values.
 
 ## 3.3.0