+++ CHANGELOG.md
@@ -1,13 +1,14 @@
 # Changelog
 
-## 3.4.0
+## 3.4.0-dev.1
 
 **Allowing custom type codecs**:
 
 - `Codec` interface is used for encoding/decoding value by type OIDs or Dart values.
 - `Codec.encode` and `Codec.decode` gets a reference to `CodecContext` which provides
    access to `encoding`, observed runtime parameters and the `TypeRegistry`.
-- `EncoderFn` value converter for generic Dart object -> Postgres-encoded bytes (for values where type is not specified).
+- `EncoderFn` value converter for generic Dart object -> Postgres-encoded bytes
+  (for values where type is not specified).
 - `RelationTracker` tracks information about relations (currently limited to `RelationMessage` caching).
 - `RuntimeParameters` to access server-provided parameter status values.